Set up script to install built-in extension node_modules, too
